перекинул работу с файлами из presenter в model, ПЗ сделано
